Implement IContainer#CreateFirewallRule
Also get rid of BlockAllOutBoundConnections. It turns out that windows
apply firewall rules in a crappy order (block rules take precendence
over allow rules). The new approach is to set the default behavior of
the firewall to block all outgoing connections and IronFrame will add
rules as needed when CreateFirewallRule is called.