Enable running containerizer as SYSTEM user
The C# Process object uses CreateProcessWithLogonW to create a process
as another user. Unfortunately this does not work as the SYSTEM user.

Given a number of frustrations we've had with running as a dedicated
administrator account (complex passwords, domain joined admins, etc.),
we decided to remove this requirement.

We now use CreateProcessAsUser to spawn the IronFrame.Host via a new
ImpersonationProcessRunner. CreateProcessAsUser may be used by the
SYSTEM user, as well as an administrator who has been granted "Replace a
process level token" policy.

Note that the IronFrame.Host job object is no longer initialized with a
name. This is due to the fact that named job objects require access to
the BaseNamedObjects directory for the current session. Under normal
circumstances, the Everyone group does not have access to this directory
for the current interactive session. Given the name is not needed, it
was removed.

[#114826119]

Signed-off-by: Ben Moss <bmoss@pivotal.io>
Signed-off-by: Amin Jamali <ajamali@pivotal.io>